desentralised-voting/requirements.txt
pycryptodomex==3.16.0
requests==2.24.0
schedule==1.1.0
